requirements.txt
pygltflib==1.16.5
pypdf>=4.1.0
Pillow>=10.0.0
Flask>=3.0.0
Flask-SQLAlchemy>=3.0.0
Flask-Login>=0.6.0
Flask-WTF>=1.2.0
Flask-Migrate>=4.0.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
gunicorn>=21.0.0
flask-cors>=4.0.0
requests>=2.31.0
PyJWT
Flask-Mail>=0.10.0
scikit-learn>=1.3.0
numpy>=1.24.0
pytest>=8.0.0
